+++ src/vsg/utils/ShaderCompiler.cpp
@@ -15,14 +15,14 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/raytracing/RayTracingPipeline.h>
 #include <vsg/state/ComputePipeline.h>
 #include <vsg/state/GraphicsPipeline.h>
-#include <vsg/vk/ShaderCompiler.h>
+#include <vsg/utils/ShaderCompiler.h>
 
 #ifdef HAS_GLSLANG
 #    include <glslang/Public/ShaderLang.h>
 #    include <glslang/SPIRV/GlslangToSpv.h>
 
-#    include "ResourceLimits.cpp"
 #    include "ResourceLimits.h"
+#    include "ResourceLimits.cpp"
 #endif
 
 #include <algorithm>
